+++ src/lib/util/cachedUserIDs.ts
@@ -2,10 +2,10 @@ import { objectEntries } from '@oldschoolgg/toolkit';
 import { Prisma } from '@prisma/client';
 import { ChannelType } from 'discord.js';
 
+import { CACHED_ACTIVE_USER_IDS } from '@/lib/cache.js';
 import { globalConfig } from '@/lib/constants.js';
 import { runTimedLoggedFn } from '@/lib/util.js';
 
-export const CACHED_ACTIVE_USER_IDS = new Set();
 CACHED_ACTIVE_USER_IDS.add(globalConfig.clientID);
 
 export const syncActiveUserIDs = async () => {
