+++ tests/unit/setup.ts
@@ -1,8 +1,6 @@
 import { vi } from 'vitest';
 import '../../src/lib/data/itemAliases.js';
 
-vi.mock('../../src/lib/postgres.js', { spy: true });
-
 vi.mock('../../src/lib/workers/index.js', { spy: true });
 
 // @ts-expect-error
